rollup_spec/requirements.txt
ethereum-execution @ git+https://github.com/ethereum/execution-specs.git
